+++ src/annbatch/samplers/__init__.py
@@ -1,10 +1,10 @@
-from ._chunk_sampler import ChunkSampler, DistributedChunkSampler
+from ._chunk_sampler import ChunkSampler, DistributedRandomSampler
 from ._random_sampler import RandomSampler
 from ._sequential_sampler import SequentialSampler
 
 __all__ = [
     "ChunkSampler",
-    "DistributedChunkSampler",
+    "DistributedRandomSampler",
     "RandomSampler",
     "SequentialSampler",
 ]