+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning][].
 
 Features:
 - Added {class}`annbatch.RandomSampler` and {class}`annbatch.SequentialSampler` as replacements for {class}`annbatch.ChunkSampler`.
+- Exposed {class}`annbatch.DistributedChunkSampler` for distributed training.
 
 Breaking:
 - Deprecated {class}`annbatch.ChunkSampler` in favor of {class}`annbatch.RandomSampler` and {class}`annbatch.SequentialSampler`.